Show name of the user who created a Payroll Run
I've also moved the `checked_by` method from the claims helper to a generic
helper that can be used in other places, as well as adding some unit tests to
the helper itself.
