+++ app/views/admin/payroll_runs/show.html.erb
@@ -25,6 +25,15 @@
       </div>
       <div class="govuk-summary-list__row">
         <dt class="govuk-summary-list__key">
+          Created by
+        </dt>
+
+        <dd class="govuk-summary-list__value">
+          <%= user_details(@payroll_run.created_by) %>
+        </dd>
+      </div>
+      <div class="govuk-summary-list__row">
+        <dt class="govuk-summary-list__key">
           Downloaded
         </dt>
